% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjust_exons.R
\name{adjust_exons_of_long_reads}
\alias{adjust_exons_of_long_reads}
\title{Adjust internal exons of full-length RNA-seq reads by majority vote}
\usage{
adjust_exons_of_long_reads(long_reads, max_exon_diff = 10)
}
\arguments{
\item{long_reads}{\code{GRangesList} object.}

\item{max_exon_diff}{Positive integer.}
}
\value{
\code{GRangesList} object with the same order of elements and same mcols as the input object.
}
\description{
Adjust internal exons of full-length RNA-seq reads by majority vote
}
\section{Details}{

The input \code{GRangesList} object is returned by the \code{extend_long_reads_to_TSS_and_PAS()} function.\cr
5'- and 3'-borders of exons in \code{long_reads} are clustered by similarity (within \code{max_exon_diff} bp).
Then the exon borders are changed to the median value of genomic coordinates observed within the cluster.
This suppresses the "fuzziness" of subalignments which are often observed in full-length RNA-seq reads.
}

