% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Call_tag_clusters.R
\name{call_TCs}
\alias{call_TCs}
\title{Call TSS or PAS sites as clusters of reads from 5'- or 3'-tag sequencing data}
\usage{
call_TCs(
  data,
  min_support = length(data),
  min_tpm = 0.05,
  max_gap = 50,
  q_trim = 0.05,
  min_score = 0.25
)
}
\arguments{
\item{data}{\code{GRangesList} object with all elements having the "score" mcols.}

\item{min_support}{Integer >= 1 (but not exceeding the number of replicates in \code{data})}

\item{min_tpm}{Non-negative numeric.}

\item{max_gap}{Non-negative integer.}

\item{q_trim}{Numeric in the range (0, 1), or NULL.}

\item{min_score}{Non-negative numeric.}
}
\value{
\code{GRanges} object containing coordinates of called tag clusters.
}
\description{
Call TSS or PAS sites as clusters of reads from 5'- or 3'-tag sequencing data
}
\section{Details}{

The input \code{data} is returned by the \code{load_BAM_files()} function.\cr
Only genomic positions with non-zero signal in at least \code{min_support} replicates are considered for TSS/PAS calling.\cr
After merging the replicates, only genomic positions with at least \code{min_tpm} signal are considered as positions with valid signal.\cr
Genomic positions with valid signal separated by not more than \code{max_gap} bases are merged into tag cluster (TC) candidates.\cr
TC candidates are trimmed to retain only genomic positions containing \code{1 - q_trim} fraction of the signal.\cr
Only TCs with TPM (averaged among replicates) exceeding \code{min_score} are considered valid TSS/PAS.
}

