% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_nascent_data.R
\name{call_transcribed_intervals}
\alias{call_transcribed_intervals}
\title{Call continuous intervals of nascent transcription from NET-seq or GRO-seq data}
\usage{
call_transcribed_intervals(
  cov,
  min_signal = 3,
  max_gapwidth = 250,
  min_width = 500
)
}
\arguments{
\item{cov}{\code{GRanges} object.}

\item{min_signal}{Positive integer.}

\item{max_gapwidth}{Non-negative integer.}

\item{min_width}{Positive integer.}
}
\value{
List of length 2:
\enumerate{
\item \code{GRanges} object containing merged intervals of nascent transcription.
\item \code{GRanges} object containing low coverage sub-intervals (gaps) within the merged transcribed intervals.
}
}
\description{
Call continuous intervals of nascent transcription from NET-seq or GRO-seq data
}
\section{Details}{

The input file is expected to be returned by the \code{load_BAM_files(mode = "nascent")} function. It contains sequencing coverage
of the chromosomes by nascent RNA-seq reads. Continuous intervals of coverage (with values not less than \code{min_signal})
are merged if the distance between them does not exceed \code{max_gapwidth} bp.
Finally, the merged intervals are filtered to have length not less than \code{min_width}.
}

