% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Call_transcript.R
\name{call_transcripts_and_genes}
\alias{call_transcripts_and_genes}
\title{Call transcript and gene models from corrected full-length RNA-seq reads}
\usage{
call_transcripts_and_genes(
  long_reads,
  skip_minor_tx = 0.01,
  max_overlap_called = 0.1,
  min_read_width = 1000,
  min_overlap_fusion = 0.5,
  clust_threshold = 0.8
)
}
\arguments{
\item{long_reads}{\code{GRangesList} object.}

\item{skip_minor_tx}{Numeric in the range (0, 1), or NULL.}

\item{max_overlap_called}{Numeric in the range [0, 1).}

\item{min_read_width}{Positive integer.}

\item{min_overlap_fusion}{Numeric in the range (0, 1].}

\item{clust_threshold}{Numeric in the range (0, 1].}
}
\value{
List of length 5:
\enumerate{
\item \code{GRanges} object (HC, MC and LC genes);
\item \code{GRangesList} object (HC, MC and LC transcripts);
\item \code{GRanges} object (fusion genes);
\item \code{GRangesList} object (fusion transcripts);
\item \code{GRangesList} object (unused long reads outside of the called genes and transcripts).
}
}
\description{
Call transcript and gene models from corrected full-length RNA-seq reads
}
\section{Details}{

The input \code{GRangesList} object is returned by the \code{detect_alignment_errors()} function.\cr
Long reads either marked as truncated by \code{extend_long_reads_to_TSS_and_PAS()}, or containing a misaligned exon (as revealed by \code{detect_alignment_errors()}), are skipped from the transcript calling procedure.
The remaining long reads are collapsed into transcripts. The transcripts are classified into high confidence (HC), medium confidence (MC) and low confidence (LC) groups:
\itemize{
\item HC transcripts are called from reads which start in a TSS and end in a PAS;
\item MC transcripts are called from TSS-only or PAS-only reads which do not overlap with any HC transcript by more than \code{max_overlap_called} fraction of either read or transcript length;
\item LC transcripts are called from reads which neither start in a TSS nor end in a PAS, and do not overlap with any HC or MC transcript by mode than \code{max_overlap_called}.
}
This iterative procedure of transcript calling ensures that highly expressed HC loci are not contaminated with less reliable MC or LC transcripts.
The MC/LC transcripts are not guaranteed to be full-length. To decrease the risk of picking up products of partial RNA degradation, MC and LC transcripts
can be called only from reads longer than \code{min_read_length} bp.\cr
The called HC, MC and LC transcripts are clustered into HC, MC and LC genes, respectively. A pair of transcripts of the same type having overlap (intersect/union) above the \code{clust_threshold}
are considered belonging to the same gene.\cr
Within each gene, the minor transcripts (collectively representing up to \code{skip_minor_tx} fraction of the reads) are skipped from further consideration. To suppress this behavior, set \code{skip_minor_tx = NULL}.\cr
Finally, transcripts which overlap at least two other disjoint transcripts by at least \code{min_overlap_fusion} fraction of their lengths, are considered fusion transcripts.
}

