% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find_alignment_errors.R
\name{detect_alignment_errors}
\alias{detect_alignment_errors}
\title{Detect the most common type of alignment errors in full-length RNA-seq reads aligned by Minimap2}
\usage{
detect_alignment_errors(long_reads, min_read_support = 2)
}
\arguments{
\item{long_reads}{\code{GRangesList} object.}

\item{min_read_support}{Positive integer.}
}
\value{
\code{GRangesList} object which is identical to the input object, except for the extra logical column \code{aln_error}.
}
\description{
Detect the most common type of alignment errors in full-length RNA-seq reads aligned by Minimap2
}
\section{Details}{

The input \code{GRangesList} object is returned by the \code{adjust_exons_of_long_reads()} function.\cr
The Minimap2 aligner is prone to under-split the full-length RNA-seq reads. In certain reads, two adjacent exons may appear as a single subalignment.
Assuming that most of the reads are still aligned correctly, one can detect alignment errors by comparing each subalignment to the set of constitutive exons.
The constitutive exons are defined here as exonic intervals supported by more than 50\% of reads in given locus (but not less than \code{min_read_support} reads).
Subalignment with an alternative 5'- and/or 3-border (relative to a constitutive exon) are considered valid alternative exons, only if the next constitutive exon is also present in the read.
Otherwise, the alignment is marked as a possible alignment error.
}

