% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_functions.R
\name{flip_strand_info}
\alias{flip_strand_info}
\title{Flip strand orientation of a GenomicRanges object}
\usage{
flip_strand_info(gr)
}
\arguments{
\item{gr}{\code{GRanges} or \code{GRangesList} object}
}
\value{
Object of the same class as input with positive and negative strands flipped
}
\description{
Flip strand orientation of a GenomicRanges object
}
