% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_BAM_files.R
\name{load_BAM_files}
\alias{load_BAM_files}
\title{Batch load BAM files}
\usage{
load_BAM_files(
  bamfiles,
  mode,
  ngs_mode = "SE",
  skip_split_aln = TRUE,
  strand_mode = 2,
  filter_tlen = 500
)
}
\arguments{
\item{bamfiles}{Character vector of BAM filenames.}

\item{mode}{Character. Must be one of the following: "long_read", "tss", "pas" or "nascent".}

\item{ngs_mode}{Character. Must be either "SE" or "PE".}

\item{skip_split_aln}{Logical.}

\item{strand_mode}{Integer. Must be either 1 or 2.}

\item{filter_tlen}{Non-negative integer.}
}
\value{
A \code{GRangesList} object, if \code{mode \%in\% c("long_reads", "tss", "pas"))}. A \code{GRanges} object, if \code{mode == "nascent")}.
}
\description{
Batch load BAM files
}
\section{Details}{

The values of \code{mode} correspond to different NGS library preparation protocols:
\itemize{
\item \code{"tss"} for 5' tag sequencing (e.g. CAGE)
\item \code{"pas"} for 3' tag sequencing (e.g. PAT-seq)
\item \code{"long_read"} for full-length RNA sequencing (e.g. Direct RNA-seq by Oxford Nanopore)
\item \code{"nascent"} for nascent RNA sequencing (e.g. NET-seq of GRO-seq)
}
if \code{mode == "long_reads"}, then each \code{GRanges} element in the returned \code{GRangesList} object corresponds to an individual long read
(each genomic range represents an exonic subalignment). The replicate samples are pooled together;\cr
If \code{mode \%in\% c("tss", "pas")}, then each \code{GRanges} element in the returned \code{GRangesList} corresponds to an individual sample (replicates are not pooled).
The input short reads are truncated to the first sequenced base and converted to genomic coverage;\cr
If \code{mode == "nascent"}, then the whole reads are converted to genomic coverage, and the replicates are pooled.\cr
The following arguments are used only \code{mode == "nascent"}:
\code{ngs_mode} describes if the sequencing was single-end or paired-end. Most often nascent RNA is sequenced in single-end mode, however some protocols (pNET-seq, plaNET-seq) require PE sequencing.\cr
\code{skip_split_aln} set to \code{TRUE} requires to skip split alignments (because nascent RNA molecules usually are not expected to be spliced).
However, if nascent RNA is sequenced by a 3rd generation platform and produces single-end long reads (e.g. chrRNA-seq), this argument has to be set to \code{FALSE}.\cr
\code{strand_mode} depends on strand orientation of the library preparation protocol. This argument is used only if \code{ngs_mode == "PE"}.\cr
\code{filter_tlen} requires to skip paired-end alignments with template length (9th field in SAM/BAM files) below the value provided. Most nascent RNA-seq protocols include an RNA fragmentation step,
thus unrealistically long inserts most probably originate either from sample contamination with mature RNA species, or from alignment errors.
}

