% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_functions.R
\name{merge_GRanges}
\alias{merge_GRanges}
\title{Merge coverage of GRanges objects}
\usage{
merge_GRanges(gr_list)
}
\arguments{
\item{gr_list}{List of \code{GRanges} objects with \code{score} metadata column}
}
\value{
\code{GRanges} object
}
\description{
Merge coverage of GRanges objects
}
