% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Refine_by_annotation.R
\name{refine_transcripts_by_annotation}
\alias{refine_transcripts_by_annotation}
\title{Refine called transcripts by an existing transcript model}
\usage{
refine_transcripts_by_annotation(
  hml_tx,
  annot_exons,
  tss,
  pas,
  fusion_tx = GenomicRanges::GRangesList(),
  max_exon_diff = 10,
  tx_flanks_up = c(-100, 100),
  tx_flanks_down = c(-100, 100),
  min_score_2 = 5,
  min_tx_cov = 0.95,
  clust_threshold = 0.8,
  min_overlap_fusion = 0.5
)
}
\arguments{
\item{hml_tx, annot_exons, fusion_tx}{\code{GRangesList} objects.}

\item{tss, pas}{\code{GRanges} objects.}

\item{max_exon_diff}{Positive integer.}

\item{tx_flanks_up, tx_flanks_down}{Integer vectors of length 2.}

\item{min_score_2}{Non-negative numeric.}

\item{min_tx_cov}{Numeric in the range (0, 1].}

\item{clust_threshold}{Numeric in the range (0, 1].}

\item{min_overlap_fusion}{Numeric in the range (0, 1].}
}
\value{
List of length 4:
\enumerate{
\item \code{GRanges} object (updated HC, MC and LC genes);
\item \code{GRangesList} object (refined HC, MC and LC transcripts);
\item \code{GRanges} object (updated fusion genes);
\item \code{GRangesList} object (updated fusion transcripts);
}
}
\description{
Refine called transcripts by an existing transcript model
}
\section{Details}{

\code{hml_tx} is the called transcript model (the second element in the list returned by the \code{call_transcripts_and_genes()} function).
\code{fusion_tx} is the called set of fusion transcripts (the fourth element in the list returned by the \code{call_transcripts_and_genes()} function).
\code{annot_exons} is a known transcript model (returned by e.g. \code{exonsBy(txdb, by = "tx")}, where \code{txdb} is a \code{GenomicFeatures} object).\cr
The function aims to adjust the called transcripts by the annotated transcripts:
\itemize{
\item 5'- and 3'-borders of called exons are adjusted to the most similar border of an annotated exon (by not more than \code{max_exon_diff} bp);
\item Annotated transcripts are classified into valid and non-valid. A valid known transcript must overlap with called TSS and PAS (both having scores above \code{min_score_2})
within \code{tx_flanks_up} and \code{tx_flanks_down} bp windows around its start and end, respectively;
\item 5'- and/or 3'-borders of called MC and LC transcripts lacking overlap with TSS and/or PAS are adjusted to the borders of the most similar mate
among the valid annotated transcripts (given that at least \code{min_tx_cov} fraction of the called transcript is covered by the annotated mate);
\item Valid annotated transcripts which do not overlap with any called transcript are copied from the annotation to the called HC transcript set;
}
In addition, the set of fusion transcripts is updated by finding called transcripts which overlap at least two valid annotated transcripts
(or an annotated and a called transcript) by at least \code{min_overlap_fusion} fraction of their lengths.
}

