% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_functions.R
\name{save_GRanges_as_bedGraph}
\alias{save_GRanges_as_bedGraph}
\title{Save stranded GRanges object as bedGraph file with strand info encoded in the sign of values}
\usage{
save_GRanges_as_bedGraph(gr, filepath, colname = "score")
}
\arguments{
\item{gr}{\code{GRanges} object with a numeric column in mcols.}

\item{filepath}{Valid path to the output file.}

\item{colname}{Character (must match the name of the numeric column).}
}
\description{
Save stranded GRanges object as bedGraph file with strand info encoded in the sign of values
}
\section{Details}{

This function saves a gzipped bedGraph file which contains both positive and negative values
(to visualize the coverage of forward and reverse strands in a genome browser).
}

