% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_functions.R
\name{write_grl_as_bed12}
\alias{write_grl_as_bed12}
\title{Save GRangesList object as BED12 file}
\usage{
write_grl_as_bed12(grl, filename, name_in_mcols = FALSE)
}
\arguments{
\item{grl}{\code{GRangesList} object.}

\item{filename}{Valid path to the output file.}

\item{name_in_mcols}{Logical.}
}
\description{
Save GRangesList object as BED12 file
}
\section{Details}{

This function does the same work as \code{rtracklayer::export(format = "BED")},
with the added benefit that the \code{S4Vectors::mcols(grl)$score} is properly exported.\cr
If \code{name_in_mcols == TRUE}, then the \code{S4Vectors::mcols(grl)$name} is exported as range names (otherwise the \code{names(grl)} is used).
}

